Do not schedule session expiry notification for 5.24+
