+++ app/init/global_event_handler.js
@@ -8,6 +8,13 @@ import DeviceInfo from 'react-native-device-info';
 import RNFetchBlob from 'rn-fetch-blob';
 import semver from 'semver/preload';
 
+import {setDeviceDimensions, setDeviceOrientation, setDeviceAsTablet, setStatusBarHeight} from '@actions/device';
+import {selectDefaultChannel} from '@actions/views/channel';
+import {showOverlay} from '@actions/navigation';
+import {loadConfigAndLicense, setDeepLinkURL, startDataCleanup} from '@actions/views/root';
+import {loadMe, logout} from '@actions/views/user';
+import {NavigationTypes, ViewTypes} from '@constants';
+import {getTranslations, resetMomentLocale} from '@i18n';
 import {setAppState, setServerVersion} from '@mm-redux/actions/general';
 import {autoUpdateTimezone} from '@mm-redux/actions/timezone';
 import {close as closeWebSocket} from '@actions/websocket';
@@ -17,15 +24,7 @@ import {getCurrentChannelId} from '@mm-redux/selectors/entities/channels';
 import {getCurrentUserId, getUser} from '@mm-redux/selectors/entities/users';
 import {isTimezoneEnabled} from '@mm-redux/selectors/entities/timezone';
 import EventEmitter from '@mm-redux/utils/event_emitter';
-
-import {setDeviceDimensions, setDeviceOrientation, setDeviceAsTablet, setStatusBarHeight} from '@actions/device';
-import {selectDefaultChannel} from '@actions/views/channel';
-import {showOverlay} from '@actions/navigation';
-import {loadConfigAndLicense, setDeepLinkURL, startDataCleanup} from '@actions/views/root';
-import {loadMe, logout} from '@actions/views/user';
-import {NavigationTypes, ViewTypes} from '@constants';
-import {getTranslations, resetMomentLocale} from '@i18n';
-import PushNotifications from 'app/push_notifications';
+import {isMinimumServerVersion} from '@mm-redux/utils/helpers';
 import {getCurrentLocale} from '@selectors/i18n';
 import initialState from '@store/initial_state';
 import Store from '@store/store';
@@ -35,6 +34,7 @@ import {getDeviceTimezoneAsync} from '@utils/timezone';
 
 import mattermostBucket from 'app/mattermost_bucket';
 import mattermostManaged from 'app/mattermost_managed';
+import PushNotifications from 'app/push_notifications';
 import LocalConfig from 'assets/config';
 
 import {getAppCredentials, removeAppCredentials} from './credentials';
@@ -150,14 +150,9 @@ class GlobalEventHandler {
         emmProvider.handleManagedConfig(true);
     };
 
-    onServerConfigChanged = (config) => {
-        this.configureAnalytics(config);
-    };
-
     onLogout = async () => {
         Store.redux.dispatch(closeWebSocket(false));
         Store.redux.dispatch(setServerVersion(''));
-        await this.resetState();
         removeAppCredentials();
         deleteFileCache();
         resetMomentLocale();
@@ -187,6 +182,12 @@ class GlobalEventHandler {
         if (this.launchApp) {
             this.launchApp();
         }
+
+        // Reset the state after sending
+        // the user to the Select server URL screen
+        // To avoid unavailable data crashes while components are
+        // still mounted.
+        this.resetState();
     };
 
     onOrientationChange = (dimensions) => {
@@ -235,6 +236,14 @@ class GlobalEventHandler {
         }
     };
 
+    onServerConfigChanged = (config) => {
+        this.configureAnalytics(config);
+
+        if (isMinimumServerVersion(Client4.serverVersion, 5, 24) && config.ExtendSessionLengthWithActivity === 'true') {
+            PushNotifications.cancelAllLocalNotifications();
+        }
+    };
+
     onServerVersionChanged = async (serverVersion) => {
         const {dispatch, getState} = Store.redux;
         const state = getState();
@@ -242,7 +251,6 @@ class GlobalEventHandler {
         const version = match && match[0];
         const locale = getCurrentLocale(state);
         const translations = getTranslations(locale);
-
         if (serverVersion) {
             if (semver.valid(version) && semver.lt(version, LocalConfig.MinServerVersion)) {
                 Alert.alert(
@@ -257,7 +265,7 @@ class GlobalEventHandler {
             } else if (state.entities.users && state.entities.users.currentUserId) {
                 dispatch(setServerVersion(serverVersion));
                 const data = await dispatch(loadConfigAndLicense());
-                this.configureAnalytics(data.config);
+                this.onServerConfigChanged(data.config);
             }
         }
     };
