+++ app/screens/select_server/index.js
@@ -4,15 +4,14 @@
 import {bindActionCreators} from 'redux';
 import {connect} from 'react-redux';
 
+import {setLastUpgradeCheck} from '@actions/views/client_upgrade';
+import {loadConfigAndLicense} from '@actions/views/root';
+import {handleServerUrlChanged} from '@actions/views/select_server';
+import {scheduleExpiredNotification} from '@actions/views/session';
 import {getPing, resetPing, setServerVersion} from '@mm-redux/actions/general';
 import {login} from '@mm-redux/actions/users';
 import {getConfig, getLicense} from '@mm-redux/selectors/entities/general';
-
-import {setLastUpgradeCheck} from 'app/actions/views/client_upgrade';
-import {handleSuccessfulLogin, scheduleExpiredNotification} from 'app/actions/views/login';
-import {loadConfigAndLicense} from 'app/actions/views/root';
-import {handleServerUrlChanged} from 'app/actions/views/select_server';
-import getClientUpgrade from 'app/selectors/client_upgrade';
+import getClientUpgrade from '@selectors/client_upgrade';
 
 import SelectServer from './select_server';
 
@@ -35,7 +34,6 @@ function mapStateToProps(state) {
 function mapDispatchToProps(dispatch) {
     return {
         actions: bindActionCreators({
-            handleSuccessfulLogin,
             getPing,
             scheduleExpiredNotification,
             handleServerUrlChanged,
