+++ app/utils/security.js
@@ -16,8 +16,8 @@ export function getMfaPreflightDone() {
 
 export function setCSRFFromCookie(url) {
     return new Promise((resolve) => {
-        CookieManager.get(url, false).then((res) => {
-            const token = res.MMCSRF;
+        CookieManager.get(url, false).then((cookies) => {
+            const token = cookies.MMCSRF;
             if (token) {
                 let value = null;
                 if (typeof token === 'object' && Object.prototype.hasOwnProperty.call(token, 'value')) {
