+++ package.json
@@ -79,6 +79,7 @@
     "reselect": "4.0.0",
     "rn-fetch-blob": "0.12.0",
     "rn-placeholder": "github:mattermost/rn-placeholder#02c629c65d0123a2eee623ada0fd17186415d3c3",
+    "scripts": "0.1.0",
     "semver": "7.1.3",
     "serialize-error": "5.0.0",
     "shallow-equals": "1.0.0",
