+++ patches/react-native-notifications+2.1.7.patch
@@ -1,15 +1,27 @@
 diff --git a/node_modules/react-native-notifications/android/app/src/main/AndroidManifest.xml b/node_modules/react-native-notifications/android/app/src/main/AndroidManifest.xml
-index 7053040..2a97b04 100644
+index 7053040..ad63eff 100644
 --- a/node_modules/react-native-notifications/android/app/src/main/AndroidManifest.xml
 +++ b/node_modules/react-native-notifications/android/app/src/main/AndroidManifest.xml
-@@ -29,6 +29,7 @@
+@@ -11,6 +11,7 @@
+         android:protectionLevel="signature" />
+     <uses-permission android:name="${applicationId}.permission.C2D_MESSAGE" />
+     <uses-permission android:name="android.permission.VIBRATE" />
++    <uses-permission android:name="android.permission.WAKE_LOCK" />
+ 
+     <application>
+ 
+@@ -29,7 +30,11 @@
  
          <service
              android:name=".fcm.FcmInstanceIdRefreshHandlerService"
 +            android:permission="android.permission.BIND_JOB_SERVICE"
              android:exported="false" />
++        <receiver android:name=".core.notification.PushNotificationPublisher"
++               android:enabled="true"
++               android:exported="false" />
      </application>
  
+ </manifest>
 diff --git a/node_modules/react-native-notifications/android/app/src/main/java/com/wix/reactnativenotifications/RNNotificationsModule.java b/node_modules/react-native-notifications/android/app/src/main/java/com/wix/reactnativenotifications/RNNotificationsModule.java
 index 7b47aed..200cca8 100644
 --- a/node_modules/react-native-notifications/android/app/src/main/java/com/wix/reactnativenotifications/RNNotificationsModule.java
